# os/bluestore: Always update the cursor position in AVL near-fit search.