@neha-ojha 2022-04-13
https://pulpito.ceph.com/?branch=wip-yuri2-testing-2022-04-13-0703 - no related failures

Failures, unrelated:
[Bug #53789: CommandFailedError (rados/test_python.sh): "RADOS object not found" causes test_rados.TestWatchNotify.test_aio_notify to fail - RADOS - Ceph](https://tracker.ceph.com/issues/53789)[](https://tracker.ceph.com/issues/55322)[](https://tracker.ceph.com/issues/55322)
[Bug #55322: test-restful.sh: mon metadata unable to be retrieved - RADOS - Ceph](https://tracker.ceph.com/issues/55322)[](https://tracker.ceph.com/issues/55323)[](https://tracker.ceph.com/issues/55323)
[Bug #55323: cephadm/test_dashboard_e2e.sh: cypress "500: Internal Server Error" caused by missing password - Dashboard - Ceph](https://tracker.ceph.com/issues/55323)

Details:
1. CommandFailedError (rados/test_python.sh): "RADOS object not found" causes test_rados.TestWatchNotify.test_aio_notify to fail - Ceph - RADOS
2. [test-restful.sh](http://test-restful.sh/): mon metadata unable to be retrieved - Ceph - RADOS
3. cephadm/test_dashboard_e2e.sh: cypress "500: Internal Server Error" caused by missing password - Ceph - Mgr - Dashboard

There was also an unrelated upgrade test failure.