@neha-ojha 2022-04-18
@markhpc I have created https://tracker.ceph.com/issues/55358 to backport the fix to older releases, I think we should include it 16.2.8.